requirements.txt
numpy
pandas
matplotlib
python-dotenv
mysql-connector-python
pymysql
scikit-learn
xgboost
catboost
mlflow
flask
cryptography